+++ src/components/App.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
-import './App.css';
 import { Layer, Rect, Stage, Group } from 'react-konva';
-import Board from './board/Board';
+import Board from './game/board/Board';
 import { ReactReduxContext, useDispatch, useSelector } from 'react-redux';
 import { useWindowDimensions } from './windowDimensions';
 import { moveBoard } from '../features/board';
 import { RootState } from '../store';
-import Players from './players/Players';
-import Sidebar from './sidebar/Sidebar';
+import Players from './game/players/Players';
+import Sidebar from './game/sidebar/Sidebar';
 
 const App = () => {
   const { width, height } = useWindowDimensions();
