+++ src/components/game/players/Players.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import { Group } from 'react-konva';
 import { useSelector } from 'react-redux';
-import { getPlayerMap } from '../../features/selectors';
-import { mapCartMap } from '../../map';
+import { getPlayerMap } from '../../../features/selectors';
+import { mapCartMap } from '../../../map';
 import { toString } from '../board/grid';
 import RoomPlayers from './RoomPlayers';
