+++ src/components/game/portal/CanvasPortal.tsx
