+++ src/components/game/sidebar/roomStack/StackRoom.tsx
@@ -1,6 +1,9 @@
 import React, { FunctionComponent } from 'react';
-import { BoundingBox } from '../../layout';
-import { Floor, StackRoom as StackRoomModel } from '../../../features/models';
+import { BoundingBox } from '../../../layout';
+import {
+  Floor,
+  StackRoom as StackRoomModel,
+} from '../../../../features/models';
 import {
   add,
   multiply,
@@ -8,10 +11,10 @@ import {
   pointsToArray,
   subtract,
   translate,
-} from '../../geometry';
+} from '../../../geometry';
 import { Group, Line, Rect, Text } from 'react-konva';
 import { useSelector } from 'react-redux';
-import { RootState } from '../../../store';
+import { RootState } from '../../../../store';
 import FlippedStackRoom from './FlippedStackRoom';
 
 interface RoofPoints {
